nvme_rdma: Destroy qpair after it is disconnected and drained
By the previous patches, a qpair is destroyed after it is actually
disconnected.

But after the qpair is destroyed, it is checked if drained by using
rqpair->current_num_sends and rqpair->current_num_recvs.

However, if the qpair is the last of a poller of a poll group,
CQ is destroyed before checking if the qpair is drained.

If CQ is destroyed, at least rqpair->current_num_recvs is not updated,
and we may get one second timeout.

This should be avoided.

Hence, destroy the qpair after it is disconnected and drained.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Change-Id: Ibd6c83e8a3e7b6e11e9b45cee42669da6d42a621
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/14278
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>